refactor(plugin): 重组插件管理与运行时模块
