+++ src-tauri/src/plugins/api/handlers.rs
@@ -1,0 +1,19 @@
+mod emitters;
+mod setters;
+mod state;
+
+pub use emitters::{
+    call_api, emit_component_event, emit_context_menu_event, emit_i18n_event, emit_log_event,
+    emit_permission_log, emit_server_ready, emit_sidebar_event, emit_ui_event,
+    register_server_log_processor,
+};
+pub use setters::{
+    set_api_call_handler, set_component_event_handler, set_context_menu_handler,
+    set_i18n_event_handler, set_log_event_handler, set_permission_log_handler,
+    set_server_ready_handler, set_sidebar_event_handler, set_ui_event_handler,
+};
+pub use state::{
+    ApiCallHandler, ComponentEventHandler, ContextMenuHandler, I18nEventHandler,
+    LogEventHandler, PermissionLogHandler, ServerReadyHandler, SidebarEventHandler,
+    UiEventHandler,
+};
