+++ src-tauri/src/plugins/loader.rs
@@ -1,10 +1,23 @@
 use crate::models::plugin::PluginManifest;
+use crate::hardcode_data::app_files::PLUGIN_MANIFEST_FILE_NAME;
 use std::fs;
 use std::path::{Path, PathBuf};
 
+/// 插件加载器
+///
+/// 负责发现插件目录、读取清单和做基础校验
 pub struct PluginLoader;
 
 impl PluginLoader {
+    /// 扫描插件目录下的可用插件文件夹
+    ///
+    /// # Parameters
+    ///
+    /// - `plugins_dir`: 插件根目录
+    ///
+    /// # Returns
+    ///
+    /// 返回所有包含 `manifest.json` 的插件目录
     pub fn discover_plugins(plugins_dir: &Path) -> Result<Vec<PathBuf>, String> {
         if !plugins_dir.exists() {
             return Ok(Vec::new());
@@ -19,7 +32,7 @@ impl PluginLoader {
             let entry = entry.map_err(|e| format!("Failed to read directory entry: {}", e))?;
             let path = entry.path();
 
-            if path.is_dir() && path.join("manifest.json").exists() {
+            if path.is_dir() && path.join(PLUGIN_MANIFEST_FILE_NAME).exists() {
                 plugin_dirs.push(path);
             }
         }
@@ -27,8 +40,17 @@ impl PluginLoader {
         Ok(plugin_dirs)
     }
 
+    /// 读取单个插件目录里的清单文件
+    ///
+    /// # Parameters
+    ///
+    /// - `plugin_dir`: 插件目录
+    ///
+    /// # Returns
+    ///
+    /// 读取成功时返回解析后的插件清单
     pub fn load_manifest(plugin_dir: &Path) -> Result<PluginManifest, String> {
-        let manifest_path = plugin_dir.join("manifest.json");
+        let manifest_path = plugin_dir.join(PLUGIN_MANIFEST_FILE_NAME);
 
         let content = fs::read_to_string(&manifest_path).map_err(|e| {
             format!("Failed to read manifest at {}: {}", manifest_path.display(), e)
@@ -41,6 +63,15 @@ impl PluginLoader {
         Ok(manifest)
     }
 
+    /// 校验插件清单的基础合法性
+    ///
+    /// # Parameters
+    ///
+    /// - `manifest`: 待校验的插件清单
+    ///
+    /// # Returns
+    ///
+    /// 校验通过时返回 `Ok(())`
     pub fn validate_manifest(manifest: &PluginManifest) -> Result<(), String> {
         if manifest.id.trim().is_empty() {
             return Err("Manifest field 'id' is required".into());
@@ -112,139 +143,5 @@ impl PluginLoader {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::models::plugin::PluginAuthor;
-    use std::fs;
-
-    fn make_temp_dir(name: &str) -> PathBuf {
-        let dir =
-            std::env::temp_dir().join(format!("sealantern_test_{}_{}", name, std::process::id()));
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap();
-        dir
-    }
-
-    fn sample_manifest_json() -> &'static str {
-        r#"{
-            "id": "com.example.test",
-            "name": "Test Plugin",
-            "version": "1.0.0",
-            "description": "A test plugin",
-            "author": { "name": "Tester" },
-            "main": "main.lua"
-        }"#
-    }
-
-    #[test]
-    fn test_discover_plugins_finds_valid_dirs() {
-        let tmp = make_temp_dir("discover");
-        let plugin_a = tmp.join("plugin-a");
-        fs::create_dir(&plugin_a).unwrap();
-        fs::write(plugin_a.join("manifest.json"), "{}").unwrap();
-
-        let no_manifest = tmp.join("no-manifest");
-        fs::create_dir(&no_manifest).unwrap();
-
-        let result = PluginLoader::discover_plugins(&tmp).unwrap();
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0], plugin_a);
-
-        let _ = fs::remove_dir_all(&tmp);
-    }
-
-    #[test]
-    fn test_discover_plugins_empty_on_missing_dir() {
-        let result =
-            PluginLoader::discover_plugins(Path::new("/nonexistent/sealantern_test_path")).unwrap();
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn test_load_manifest_success() {
-        let tmp = make_temp_dir("load");
-        fs::write(tmp.join("manifest.json"), sample_manifest_json()).unwrap();
-
-        let manifest = PluginLoader::load_manifest(&tmp).unwrap();
-        assert_eq!(manifest.id, "com.example.test");
-        assert_eq!(manifest.name, "Test Plugin");
-        assert_eq!(manifest.version, "1.0.0");
-        assert_eq!(manifest.main, "main.lua");
-
-        let _ = fs::remove_dir_all(&tmp);
-    }
-
-    #[test]
-    fn test_load_manifest_file_not_found() {
-        let tmp = make_temp_dir("load_missing");
-        let result = PluginLoader::load_manifest(&tmp);
-        assert!(result.is_err());
-
-        let _ = fs::remove_dir_all(&tmp);
-    }
-
-    #[test]
-    fn test_validate_manifest_ok() {
-        let manifest = PluginManifest {
-            id: "com.example.test".into(),
-            name: "Test".into(),
-            version: "1.0.0".into(),
-            description: "desc".into(),
-            author: PluginAuthor {
-                name: "Dev".into(),
-                email: None,
-                url: None,
-            },
-            main: "main.lua".into(),
-            license: None,
-            homepage: None,
-            repository: None,
-            engines: None,
-            permissions: vec![],
-            ui: None,
-            events: vec![],
-            commands: vec![],
-            dependencies: Default::default(),
-            optional_dependencies: Default::default(),
-            icon: None,
-            settings: None,
-            sidebar: None,
-            locales: None,
-            include: vec![],
-        };
-        assert!(PluginLoader::validate_manifest(&manifest).is_ok());
-    }
-
-    #[test]
-    fn test_validate_manifest_empty_id() {
-        let manifest = PluginManifest {
-            id: "".into(),
-            name: "Test".into(),
-            version: "1.0.0".into(),
-            description: "desc".into(),
-            author: PluginAuthor {
-                name: "Dev".into(),
-                email: None,
-                url: None,
-            },
-            main: "main.lua".into(),
-            license: None,
-            homepage: None,
-            repository: None,
-            engines: None,
-            permissions: vec![],
-            ui: None,
-            events: vec![],
-            commands: vec![],
-            dependencies: Default::default(),
-            optional_dependencies: Default::default(),
-            icon: None,
-            settings: None,
-            sidebar: None,
-            locales: None,
-            include: vec![],
-        };
-        let err = PluginLoader::validate_manifest(&manifest).unwrap_err();
-        assert!(err.contains("id"));
-    }
-}
+#[path = "../../tests/unit/plugins_loader_tests.rs"]
+mod tests;
