+++ src-tauri/src/plugins/manager/lifecycle/runtime.rs
@@ -1,0 +1,6 @@
+mod disable;
+mod enable;
+mod shared;
+
+pub(super) use disable::{disable_plugin, disable_plugin_internal};
+pub(super) use enable::enable_plugin;
